+++ test/cldr_test.exs
@@ -37,7 +37,8 @@ defmodule Cldr.Test do
                "date_fields",
                "dates",
                "territories",
-               "languages"
+               "languages",
+               "delimiters"
              ]
   end